feat: add unit test for sign up using jest
